test(providers): improve http test split organization and robustness
Redistribute non-auth tests from auth.test.ts to appropriate files:
- Move urlEncodeRawRequestPath, extractBodyFromRawRequest, Body file
  resolution, and Sanitization blocks to core.test.ts
- Move createSessionParser, Token Estimation unit, and Abort Signal
  Handling blocks to request.test.ts
- auth.test.ts now contains only authentication tests

Harden test infrastructure:
- fetchWithCache default mock throws instead of resolving undefined
- importModule mock throws for unknown function names
- Add beforeEach to reset cliState.config (fixes order-dependent flakiness)
- Replace empty .catch blocks with error assertions in OAuth tests
- Add explicit toHaveBeenCalled checks before accessing mock.calls

Add file-level scope comments to all five test files.
