+++ test/providers/http/auth.test.ts
@@ -1,3 +1,4 @@
+// Authentication tests: RSA signatures, JKS keystores, OAuth token refresh, file-based auth, bearer/API key auth.
 import './setup';
 
 import crypto from 'crypto';
@@ -9,17 +10,9 @@ import { afterEach, beforeEach, describe, expect, it, type MockInstance, vi } fr
 import { fetchWithCache } from '../../../src/cache';
 import { importModule } from '../../../src/esm';
 import logger from '../../../src/logger';
-import {
-  createSessionParser,
-  estimateTokenCount,
-  extractBodyFromRawRequest,
-  HttpProvider,
-  urlEncodeRawRequestPath,
-} from '../../../src/providers/http';
+import { HttpProvider } from '../../../src/providers/http';
 import { runPython } from '../../../src/python/pythonUtils';
-import { maybeLoadConfigFromExternalFile } from '../../../src/util/file';
 import { TOKEN_REFRESH_BUFFER_MS } from '../../../src/util/oauth';
-import { sanitizeObject, sanitizeUrl } from '../../../src/util/sanitizer';
 import { createDeferred, mockProcessEnv } from '../../util/utils';
 
 describe('RSA signature authentication', () => {
@@ -402,1185 +395,6 @@ describe('RSA signature authentication', () => {
   });
 });
 
-describe('createSessionParser', () => {
-  it('should return empty string when no parser is provided', async () => {
-    const parser = await createSessionParser(undefined);
-    const result = parser({ headers: {}, body: {} });
-    expect(result).toBe('');
-  });
-
-  it('should handle function parser', async () => {
-    const functionParser = ({ headers }: { headers: Record<string, string> }) =>
-      headers['session-id'];
-    const parser = await createSessionParser(functionParser);
-    const result = parser({ headers: { 'session-id': 'test-session' } });
-    expect(result).toBe('test-session');
-  });
-
-  it('should handle header path expression', async () => {
-    const parser = await createSessionParser('data.headers["x-session-id"]');
-    const result = parser({
-      headers: { 'x-session-id': 'test-session' },
-      body: {},
-    });
-    expect(result).toBe('test-session');
-  });
-
-  it('should handle body path expression', async () => {
-    const parser = await createSessionParser('data.body.session.id');
-    const result = parser({
-      headers: {},
-      body: { session: { id: 'test-session' } },
-    });
-    expect(result).toBe('test-session');
-  });
-
-  it('should handle file:// parser', async () => {
-    const mockParser = vi.fn(({ headers }) => headers['session-id']);
-    vi.mocked(importModule).mockResolvedValueOnce(mockParser);
-
-    const parser = await createSessionParser('file://session-parser.js');
-    const result = parser({ headers: { 'session-id': 'test-session' } });
-
-    expect(result).toBe('test-session');
-    expect(importModule).toHaveBeenCalledWith(
-      path.resolve('/mock/base/path', 'session-parser.js'),
-      undefined,
-    );
-  });
-
-  it('should handle file:// parser with specific function', async () => {
-    const mockParser = vi.fn(({ body }) => body.sessionId);
-    vi.mocked(importModule).mockResolvedValueOnce(mockParser);
-
-    const parser = await createSessionParser('file://session-parser.js:parseSession');
-    const result = parser({ headers: {}, body: { sessionId: 'test-session' } });
-
-    expect(result).toBe('test-session');
-    expect(importModule).toHaveBeenCalledWith(
-      path.resolve('/mock/base/path', 'session-parser.js'),
-      'parseSession',
-    );
-  });
-
-  it('should throw error for malformed file:// parser', async () => {
-    vi.mocked(importModule).mockResolvedValueOnce({});
-
-    await expect(createSessionParser('file://invalid-parser.js')).rejects.toThrow(
-      /Response transform malformed/,
-    );
-  });
-
-  it('should handle complex body path expression', async () => {
-    const parser = await createSessionParser('data.body.data.attributes.session.id');
-    const result = parser({
-      headers: {},
-      body: {
-        data: {
-          attributes: {
-            session: {
-              id: 'test-session',
-            },
-          },
-        },
-      },
-    });
-    expect(result).toBe('test-session');
-  });
-});
-
-describe('urlEncodeRawRequestPath', () => {
-  it('should not modify request with no query parameters', () => {
-    const rawRequest = 'GET /api/data HTTP/1.1';
-    const result = urlEncodeRawRequestPath(rawRequest);
-    expect(result).toBe(rawRequest);
-  });
-
-  it('should not modify request with simple query parameters', () => {
-    const rawRequest = 'GET /api/data?key=value HTTP/1.1';
-    const result = urlEncodeRawRequestPath(rawRequest);
-    expect(result).toBe(rawRequest);
-  });
-
-  it('should encode URL with spaces in query parameters', () => {
-    const rawRequest = 'GET /api/data?query=hello world HTTP/1.1';
-    const result = urlEncodeRawRequestPath(rawRequest);
-    expect(result).toBe('GET /api/data?query=hello%20world HTTP/1.1');
-  });
-
-  it('should encode URL with already percent-encoded characters', () => {
-    const rawRequest = 'GET /api/data?query=already%20encoded HTTP/1.1';
-    const result = urlEncodeRawRequestPath(rawRequest);
-    expect(result).toBe('GET /api/data?query=already%20encoded HTTP/1.1');
-  });
-
-  it('should throw error when modifying malformed request with no URL', () => {
-    const rawRequest = 'GET HTTP/1.1';
-    expect(() => urlEncodeRawRequestPath(rawRequest)).toThrow(/not valid/);
-  });
-
-  it('should handle complete raw request with headers', () => {
-    const rawRequest = dedent`
-      GET /summarized?topic=hello world&start=01/01/2025&end=01/07/2025&auto_extract_keywords=false HTTP/2
-      Host: foo.bar.com
-      User-Agent: curl/8.7.1
-      Accept: application/json
-    `;
-    const expected = dedent`
-      GET /summarized?topic=hello%20world&start=01/01/2025&end=01/07/2025&auto_extract_keywords=false HTTP/2
-      Host: foo.bar.com
-      User-Agent: curl/8.7.1
-      Accept: application/json
-    `;
-    const result = urlEncodeRawRequestPath(rawRequest);
-    expect(result).toBe(expected);
-  });
-
-  it('should handle POST request with JSON body', () => {
-    const rawRequest = dedent`
-      POST /api/submit?param=hello world HTTP/1.1
-      Host: example.com
-      Content-Type: application/json
-
-      {"key": "value with spaces", "date": "01/01/2025"}
-    `;
-    const expected = dedent`
-      POST /api/submit?param=hello%20world HTTP/1.1
-      Host: example.com
-      Content-Type: application/json
-
-      {"key": "value with spaces", "date": "01/01/2025"}
-    `;
-    const result = urlEncodeRawRequestPath(rawRequest);
-    expect(result).toBe(expected);
-  });
-
-  it('should handle URL with path containing spaces', () => {
-    const rawRequest = 'GET /path with spaces/resource HTTP/1.1';
-    const result = urlEncodeRawRequestPath(rawRequest);
-    expect(result).toBe('GET /path%20with%20spaces/resource HTTP/1.1');
-  });
-
-  it('should handle URL with special characters in path and query', () => {
-    const rawRequest = 'GET /path/with [brackets]?param=value&special=a+b+c HTTP/1.1';
-    const result = urlEncodeRawRequestPath(rawRequest);
-    expect(result).toBe('GET /path/with%20[brackets]?param=value&special=a+b+c HTTP/1.1');
-  });
-
-  it('should handle completely misformed first line', () => {
-    const rawRequest = 'This is not a valid HTTP request line';
-    expect(() => urlEncodeRawRequestPath(rawRequest)).toThrow(/not valid/);
-  });
-
-  it('should handle request with no HTTP protocol version', () => {
-    const rawRequest = 'GET /api/data?query=test';
-    expect(() => urlEncodeRawRequestPath(rawRequest)).toThrow(/not valid/);
-  });
-
-  it('should handle request with different HTTP methods', () => {
-    const methods = ['POST', 'PUT', 'DELETE', 'PATCH', 'OPTIONS', 'HEAD'];
-
-    for (const method of methods) {
-      const rawRequest = dedent`
-        ${method} /api/submit?param=hello world HTTP/1.1
-        Host: example.com
-        Content-Type: application/json
-
-      {"key": "value with spaces", "date": "01/01/2025"}
-      `;
-      const expected = dedent`
-        ${method} /api/submit?param=hello%20world HTTP/1.1
-        Host: example.com
-        Content-Type: application/json
-
-      {"key": "value with spaces", "date": "01/01/2025"}
-    `;
-      const result = urlEncodeRawRequestPath(rawRequest);
-      expect(result).toBe(expected);
-    }
-  });
-});
-
-describe('extractBodyFromRawRequest', () => {
-  it('should extract body from a simple POST request', () => {
-    const rawRequest = dedent`
-      POST /api/submit HTTP/1.1
-      Host: example.com
-      Content-Type: application/json
-
-      {"key": "value"}
-    `;
-    expect(extractBodyFromRawRequest(rawRequest)).toBe('{"key": "value"}');
-  });
-
-  it('should extract multipart/form-data body', () => {
-    const rawRequest = dedent`
-      POST /api/upload HTTP/1.1
-      Host: example.com
-      Content-Type: multipart/form-data; boundary=----Boundary123
-
-      ------Boundary123
-      Content-Disposition: form-data; name="field1"
-
-      value1
-      ------Boundary123--
-    `;
-    const body = extractBodyFromRawRequest(rawRequest);
-    expect(body).toContain('------Boundary123');
-    expect(body).toContain('Content-Disposition: form-data; name="field1"');
-    expect(body).toContain('value1');
-    expect(body).toContain('------Boundary123--');
-  });
-
-  it('should extract x-www-form-urlencoded body', () => {
-    const rawRequest = dedent`
-      POST /api/submit HTTP/1.1
-      Host: example.com
-      Content-Type: application/x-www-form-urlencoded
-
-      field1=value1&field2=value2
-    `;
-    expect(extractBodyFromRawRequest(rawRequest)).toBe('field1=value1&field2=value2');
-  });
-
-  it('should return undefined for GET request without body', () => {
-    const rawRequest = dedent`
-      GET /api/data HTTP/1.1
-      Host: example.com
-    `;
-    expect(extractBodyFromRawRequest(rawRequest)).toBeUndefined();
-  });
-
-  it('should return undefined for request with empty body', () => {
-    const rawRequest = dedent`
-      POST /api/submit HTTP/1.1
-      Host: example.com
-      Content-Type: application/json
-
-    `;
-    expect(extractBodyFromRawRequest(rawRequest)).toBeUndefined();
-  });
-
-  it('should handle body containing \\r\\n\\r\\n sequence', () => {
-    const rawRequest =
-      'POST /api/submit HTTP/1.1\r\n' +
-      'Host: example.com\r\n' +
-      'Content-Type: text/plain\r\n' +
-      '\r\n' +
-      'line1\r\n\r\nline2';
-    expect(extractBodyFromRawRequest(rawRequest)).toBe('line1\r\n\r\nline2');
-  });
-
-  it('should normalize mixed line endings', () => {
-    const rawRequest = 'POST /api/submit HTTP/1.1\nHost: example.com\r\n\r\nbody content';
-    expect(extractBodyFromRawRequest(rawRequest)).toBe('body content');
-  });
-
-  it('should trim leading and trailing whitespace from body', () => {
-    const rawRequest = dedent`
-      POST /api/submit HTTP/1.1
-      Host: example.com
-
-
-        body with whitespace
-
-    `;
-    expect(extractBodyFromRawRequest(rawRequest)).toBe('body with whitespace');
-  });
-
-  it('should handle special characters in body', () => {
-    const rawRequest = dedent`
-      POST /api/submit HTTP/1.1
-      Host: example.com
-      Content-Type: application/json
-
-      {"emoji": "🎉", "unicode": "日本語", "ampersand": "&"}
-    `;
-    expect(extractBodyFromRawRequest(rawRequest)).toBe(
-      '{"emoji": "🎉", "unicode": "日本語", "ampersand": "&"}',
-    );
-  });
-
-  it('should handle multiple headers before body', () => {
-    const rawRequest = dedent`
-      POST /api/submit HTTP/1.1
-      Host: example.com
-      Content-Type: application/json
-      Authorization: Bearer token123
-      X-Custom-Header: custom-value
-      Accept: application/json
-
-      {"data": "test"}
-    `;
-    expect(extractBodyFromRawRequest(rawRequest)).toBe('{"data": "test"}');
-  });
-});
-
-describe('Token Estimation', () => {
-  describe('estimateTokenCount', () => {
-    it('should count tokens using word-based method', () => {
-      const text = 'Hello world this is a test';
-      const result = estimateTokenCount(text, 1.3);
-      expect(result).toBe(Math.ceil(6 * 1.3)); // 6 words * 1.3 = 7.8, ceil = 8
-    });
-
-    it('should handle empty text', () => {
-      expect(estimateTokenCount('', 1.3)).toBe(0);
-      expect(estimateTokenCount(null as any, 1.3)).toBe(0);
-      expect(estimateTokenCount(undefined as any, 1.3)).toBe(0);
-    });
-
-    it('should filter out empty words', () => {
-      const text = 'hello   world    test'; // Multiple spaces
-      const result = estimateTokenCount(text, 1.0);
-      expect(result).toBe(3); // Should count 3 words, not split on every space
-    });
-
-    it('should use default multiplier when not provided', () => {
-      const text = 'hello world';
-      const result = estimateTokenCount(text);
-      expect(result).toBe(Math.ceil(2 * 1.3)); // Default multiplier is 1.3
-    });
-  });
-});
-
-describe('Body file resolution', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it('should resolve file:// references in body configuration', () => {
-    const mockTransactions = [
-      { id: '1', amount: '100.50', date: '2025-06-01' },
-      { id: '2', amount: '250.75', date: '2025-06-02' },
-    ];
-
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return {
-        query: '{{prompt}}',
-        date: '2025-06-03T22:01:13.797Z',
-        transactions: mockTransactions,
-      };
-    });
-
-    const provider = new HttpProvider('http://test.com', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: {
-          query: '{{prompt}}',
-          date: '2025-06-03T22:01:13.797Z',
-          transactions: 'file://./test_data/transactions.csv',
-        },
-      },
-    });
-
-    // Verify maybeLoadConfigFromExternalFile was called with the body
-    expect(maybeLoadConfigFromExternalFile).toHaveBeenCalledWith({
-      query: '{{prompt}}',
-      date: '2025-06-03T22:01:13.797Z',
-      transactions: 'file://./test_data/transactions.csv',
-    });
-
-    // The provider should have the resolved config
-    expect(provider['config'].body).toEqual({
-      query: '{{prompt}}',
-      date: '2025-06-03T22:01:13.797Z',
-      transactions: mockTransactions,
-    });
-  });
-
-  it('should resolve nested file:// references in body configuration', () => {
-    const mockTransactions = [
-      { id: '1', amount: '100.50' },
-      { id: '2', amount: '250.75' },
-    ];
-    const mockConfig = {
-      api_key: 'test-key-123',
-      timeout: 5000,
-    };
-    const mockUsers = [
-      { name: 'John', email: 'john@example.com' },
-      { name: 'Jane', email: 'jane@example.com' },
-    ];
-
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return {
-        query: '{{prompt}}',
-        data: {
-          transactions: mockTransactions,
-          settings: mockConfig,
-          nested: {
-            users: mockUsers,
-          },
-        },
-      };
-    });
-
-    const provider = new HttpProvider('http://test.com', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: {
-          query: '{{prompt}}',
-          data: {
-            transactions: 'file://./transactions.csv',
-            settings: 'file://./config.json',
-            nested: {
-              users: 'file://./users.csv',
-            },
-          },
-        },
-      },
-    });
-
-    // Verify the nested structure was resolved
-    expect(provider['config'].body).toEqual({
-      query: '{{prompt}}',
-      data: {
-        transactions: mockTransactions,
-        settings: mockConfig,
-        nested: {
-          users: mockUsers,
-        },
-      },
-    });
-  });
-
-  it('should resolve file:// references in arrays', () => {
-    const mockConfig = {
-      api_key: 'test-key-123',
-      timeout: 5000,
-    };
-    const mockUsers = [{ name: 'John', email: 'john@example.com' }];
-
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return [
-        'regular string',
-        mockConfig,
-        {
-          inside_array: mockUsers,
-        },
-      ];
-    });
-
-    const provider = new HttpProvider('http://test.com', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: [
-          'regular string',
-          'file://./config.json',
-          {
-            inside_array: 'file://./users.csv',
-          },
-        ],
-      },
-    });
-
-    // Verify arrays with file references were resolved
-    expect(provider['config'].body).toEqual([
-      'regular string',
-      mockConfig,
-      {
-        inside_array: mockUsers,
-      },
-    ]);
-  });
-
-  it('should not affect body when no file:// references are present', () => {
-    const originalBody = {
-      query: '{{prompt}}',
-      regular: 'data',
-      nested: {
-        value: 123,
-        array: ['a', 'b', 'c'],
-      },
-    };
-
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return originalBody;
-    });
-
-    const provider = new HttpProvider('http://test.com', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: originalBody,
-      },
-    });
-
-    // Body should remain unchanged
-    expect(provider['config'].body).toEqual(originalBody);
-  });
-
-  it('should work with string body containing file:// reference', () => {
-    const mockContent = 'This is the content from the file';
-
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return mockContent;
-    });
-
-    const provider = new HttpProvider('http://test.com', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'text/plain' },
-        body: 'file://./content.txt',
-      },
-    });
-
-    // String body should be resolved to file content
-    expect(provider['config'].body).toBe(mockContent);
-  });
-
-  it('should use resolved body in API calls', async () => {
-    const mockTransactions = [
-      { id: '1', amount: '100.50' },
-      { id: '2', amount: '250.75' },
-    ];
-
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return {
-        query: '{{prompt}}',
-        transactions: mockTransactions,
-      };
-    });
-
-    const provider = new HttpProvider('http://test.com', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: {
-          query: '{{prompt}}',
-          transactions: 'file://./transactions.csv',
-        },
-      },
-    });
-
-    const mockResponse = {
-      data: JSON.stringify({ result: 'success' }),
-      status: 200,
-      statusText: 'OK',
-      cached: false,
-    };
-    vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-    await provider.callApi('test prompt');
-
-    // Verify the fetch was called with the resolved body
-    // Note: processJsonBody may parse JSON strings, so check the actual call
-    expect(fetchWithCache).toHaveBeenCalled();
-
-    const actualCall = vi.mocked(fetchWithCache).mock.calls[0];
-    expect(actualCall).toBeDefined();
-    expect(actualCall[0]).toBe('http://test.com');
-
-    const requestOptions = actualCall[1];
-    expect(requestOptions).toBeDefined();
-    expect(requestOptions!.method).toBe('POST');
-    expect(requestOptions!.headers).toEqual({ 'content-type': 'application/json' });
-
-    // Parse the actual body to verify it contains the right data
-    const bodyStr = requestOptions!.body as string;
-    const bodyObj = JSON.parse(bodyStr);
-    expect(bodyObj.query).toBe('test prompt');
-    expect(bodyObj.transactions).toBeDefined();
-    expect(bodyObj.transactions.length).toBe(2);
-    // The transactions are there, whether as strings or numbers
-    expect(bodyObj.transactions[0].id).toBeDefined();
-    expect(bodyObj.transactions[1].id).toBeDefined();
-  });
-
-  it('should handle GET requests without body file resolution', () => {
-    // maybeLoadConfigFromExternalFile should not be called for GET requests without body
-    vi.mocked(maybeLoadConfigFromExternalFile).mockClear();
-
-    new HttpProvider('http://test.com', {
-      config: {
-        method: 'GET',
-      },
-    });
-
-    // Should not call maybeLoadConfigFromExternalFile since there's no body
-    expect(maybeLoadConfigFromExternalFile).not.toHaveBeenCalled();
-  });
-
-  it('should handle complex nested file resolutions with mixed content', () => {
-    const mockData = {
-      simple: 'value',
-      fileRef: { loaded: 'from file' },
-      nested: {
-        another: 'regular',
-        fileData: [1, 2, 3],
-        deeper: {
-          moreFiles: { data: 'loaded' },
-        },
-      },
-      arrayWithFiles: ['string', { fromFile: true }, ['nested', 'array']],
-    };
-
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return mockData;
-    });
-
-    const provider = new HttpProvider('http://test.com', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: {
-          simple: 'value',
-          fileRef: 'file://./data.json',
-          nested: {
-            another: 'regular',
-            fileData: 'file://./numbers.json',
-            deeper: {
-              moreFiles: 'file://./more.json',
-            },
-          },
-          arrayWithFiles: ['string', 'file://./object.json', ['nested', 'array']],
-        },
-      },
-    });
-
-    expect(provider['config'].body).toEqual(mockData);
-  });
-});
-
-describe('HttpProvider - Sanitization', () => {
-  const testUrl = 'http://example.com/api';
-  let loggerDebugSpy: MockInstance;
-
-  beforeEach(() => {
-    vi.clearAllMocks();
-    loggerDebugSpy = vi.spyOn(logger, 'debug');
-  });
-
-  afterEach(() => {
-    loggerDebugSpy.mockRestore();
-  });
-
-  it('should sanitize pfxPassword in debug logs', async () => {
-    const provider = new HttpProvider(testUrl, {
-      config: {
-        method: 'POST',
-        body: { test: 'value' },
-        // Don't include signatureAuth to avoid signature generation errors
-        headers: {
-          'X-Custom': 'test-header',
-        },
-      },
-    });
-
-    // Mock the sanitizeConfigForLogging function by spying on the actual config used in the log
-    const mockResponse = {
-      data: '{"result": "test"}',
-      status: 200,
-      statusText: 'OK',
-      cached: false,
-    };
-    vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-    await provider.callApi('test prompt');
-
-    // Instead of testing pfxPassword directly, let's test a working scenario
-    expect(loggerDebugSpy).toHaveBeenCalledWith(
-      expect.stringContaining('Calling http://example.com/api with config'),
-      expect.anything(),
-    );
-  });
-
-  it('should sanitize Authorization header in debug logs', async () => {
-    // Mock the file resolution to return a simple body to avoid conflicts
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return {
-        simple: 'test-value',
-      };
-    });
-
-    const provider = new HttpProvider(testUrl, {
-      config: {
-        method: 'POST',
-        body: { simple: 'test-value' },
-        headers: {
-          Authorization: 'Bearer secret-token-12345',
-          'Content-Type': 'application/json',
-        },
-      },
-    });
-
-    const mockResponse = {
-      data: '{"result": "test"}',
-      status: 200,
-      statusText: 'OK',
-      cached: false,
-    };
-    vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-    await provider.callApi('test prompt');
-
-    // Verify the logger was called (actual sanitization happens internally in logger)
-    expect(loggerDebugSpy).toHaveBeenCalledWith(
-      expect.stringContaining('Calling'),
-      expect.objectContaining({ config: expect.any(Object) }),
-    );
-  });
-
-  it('should sanitize multiple credential fields', async () => {
-    // Simplified test without signature auth to avoid certificate issues
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return {
-        simple: 'test-value',
-      };
-    });
-
-    const provider = new HttpProvider(testUrl, {
-      config: {
-        method: 'POST',
-        body: { simple: 'test-value' },
-        headers: {
-          Authorization: 'Bearer token-123',
-          'X-API-Key': 'api-key-456',
-        },
-        apiKey: 'main-api-key-789',
-        token: 'bearer-token-000',
-        password: 'config-password-111',
-      },
-    });
-
-    const mockResponse = {
-      data: '{"result": "test"}',
-      status: 200,
-      statusText: 'OK',
-      cached: false,
-    };
-    vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-    await provider.callApi('test prompt');
-
-    // Verify the logger was called (actual sanitization happens internally in logger)
-    expect(loggerDebugSpy).toHaveBeenCalledWith(
-      expect.stringContaining('Calling'),
-      expect.objectContaining({ config: expect.any(Object) }),
-    );
-  });
-
-  it('should preserve non-sensitive fields', async () => {
-    vi.mocked(maybeLoadConfigFromExternalFile).mockImplementation(function () {
-      return {
-        simple: 'test-value',
-      };
-    });
-
-    const provider = new HttpProvider(testUrl, {
-      config: {
-        method: 'POST',
-        body: { simple: 'test-value' },
-        headers: {
-          'Content-Type': 'application/json',
-          'User-Agent': 'test-agent',
-        },
-        timeout: 5000,
-        maxRetries: 3,
-      },
-    });
-
-    const mockResponse = {
-      data: '{"result": "test"}',
-      status: 200,
-      statusText: 'OK',
-      cached: false,
-    };
-    vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-    await provider.callApi('test prompt');
-
-    const debugCall = loggerDebugSpy.mock.calls.find(
-      (call: any) => call[0]?.includes('Calling') && call[0]?.includes('with config'),
-    );
-    expect(debugCall).toBeDefined();
-
-    const context = debugCall?.[1];
-    const contextStr = JSON.stringify(context);
-    expect(contextStr).toContain('"content-type":"application/json"'); // lowercase
-    expect(contextStr).toContain('"user-agent":"test-agent"'); // lowercase
-    // Note: timeout and maxRetries are not included in the rendered config that gets logged
-    expect(contextStr).not.toContain('[REDACTED]');
-  });
-
-  describe('Header sanitization in logs', () => {
-    it('should sanitize sensitive headers while preserving functionality', async () => {
-      const provider = new HttpProvider('https://api.example.com/test', {
-        config: {
-          method: 'POST',
-          body: { message: '{{ prompt }}' },
-          headers: {
-            Authorization: 'Bearer secret-token-12345',
-            'X-API-Key': 'sk-test-abc123',
-            'Content-Type': 'application/json',
-          },
-        },
-      });
-
-      const mockResponse = {
-        data: '{"success": true}',
-        status: 200,
-        statusText: 'OK',
-        cached: false,
-      };
-      vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-      await provider.callApi('test message');
-
-      // Verify the logger was called with a context object
-      expect(loggerDebugSpy).toHaveBeenCalledWith(
-        expect.stringContaining('Calling'),
-        expect.objectContaining({ config: expect.any(Object) }),
-      );
-
-      // Extract the context object that was passed to the logger
-      const logCall = loggerDebugSpy.mock.calls.find((call: any) => call[0]?.includes('Calling'));
-      expect(logCall).toBeDefined();
-      const loggedContext = logCall![1];
-
-      // Verify that when the logger sanitizes this context, sensitive headers are redacted
-      const sanitizedContext = sanitizeObject(loggedContext, { context: 'test' });
-      const sanitizedHeaders = sanitizedContext.config.headers;
-
-      // Check for headers with different casing (they may be normalized)
-      const authHeader =
-        sanitizedHeaders.Authorization ||
-        sanitizedHeaders.authorization ||
-        Object.entries(sanitizedHeaders).find(
-          ([key]) => key.toLowerCase() === 'authorization',
-        )?.[1];
-      const apiKeyHeader =
-        sanitizedHeaders['X-API-Key'] ||
-        sanitizedHeaders['x-api-key'] ||
-        Object.entries(sanitizedHeaders).find(([key]) => key.toLowerCase() === 'x-api-key')?.[1];
-
-      expect(authHeader).toEqual('[REDACTED]');
-      expect(apiKeyHeader).toEqual('[REDACTED]');
-
-      // Verify actual functionality works - fetchWithCache should get real headers
-      expect(fetchWithCache).toHaveBeenCalledWith(
-        'https://api.example.com/test',
-        expect.objectContaining({
-          headers: expect.objectContaining({
-            authorization: 'Bearer secret-token-12345', // Real token sent (lowercase)
-            'x-api-key': 'sk-test-abc123', // Real key sent
-          }),
-        }),
-        expect.any(Number),
-        'text',
-        undefined,
-        undefined,
-      );
-    });
-
-    it('should preserve non-sensitive headers in logs', async () => {
-      const provider = new HttpProvider('https://api.example.com/test', {
-        config: {
-          method: 'POST',
-          body: { message: '{{ prompt }}' },
-          headers: {
-            'Content-Type': 'application/json',
-            'User-Agent': 'test-client/1.0',
-            Accept: 'application/json',
-          },
-        },
-      });
-
-      const mockResponse = {
-        data: '{"success": true}',
-        status: 200,
-        statusText: 'OK',
-        cached: false,
-      };
-      vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-      await provider.callApi('test message');
-
-      // Verify the logger was called
-      expect(loggerDebugSpy).toHaveBeenCalledWith(
-        expect.stringContaining('Calling'),
-        expect.objectContaining({ config: expect.any(Object) }),
-      );
-    });
-  });
-
-  describe('URL sanitization', () => {
-    it('should sanitize URL query parameters', async () => {
-      const provider = new HttpProvider(
-        'https://api.example.com/test?api_key=secret123&format=json',
-        {
-          config: {
-            method: 'GET',
-          },
-        },
-      );
-
-      const mockResponse = {
-        data: '{"success": true}',
-        status: 200,
-        statusText: 'OK',
-        cached: false,
-      };
-      vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-      await provider.callApi('test');
-
-      const debugCall = loggerDebugSpy.mock.calls.find((call: any) => call[0].includes('Calling'));
-      expect(debugCall).toBeDefined();
-
-      const logMessage = debugCall![0];
-      expect(logMessage).toContain('api_key=%5BREDACTED%5D');
-      expect(logMessage).toContain('format=json'); // Non-sensitive param preserved
-      // Note: The URL in config object may contain the original secret, but the main URL is sanitized
-    });
-
-    it('should work with standalone sanitizeUrl function', () => {
-      const testCases = [
-        {
-          input: 'https://user:pass@api.com/test?api_key=secret&normal=value',
-          expectContains: ['***:***', 'api_key=%5BREDACTED%5D', 'normal=value'],
-          expectNotContains: ['user', 'secret'],
-        },
-        {
-          input: 'https://api.com/test?token=bearer123&id=123',
-          expectContains: ['token=%5BREDACTED%5D', 'id=123'],
-          expectNotContains: ['bearer123'],
-        },
-      ];
-
-      testCases.forEach(({ input, expectContains, expectNotContains }) => {
-        const result = sanitizeUrl(input);
-
-        expectContains.forEach((expectedText) => {
-          expect(result).toContain(expectedText);
-        });
-
-        expectNotContains.forEach((secretText) => {
-          expect(result).not.toContain(secretText);
-        });
-      });
-    });
-  });
-
-  describe('Combined sanitization scenarios', () => {
-    it('should handle both URL and header sanitization together', async () => {
-      const provider = new HttpProvider('https://api.example.com/test?api_key=url_secret123', {
-        config: {
-          method: 'POST',
-          body: { data: '{{ prompt }}' },
-          headers: {
-            Authorization: 'Bearer header_secret456',
-          },
-        },
-      });
-
-      const mockResponse = {
-        data: '{"result": "success"}',
-        status: 200,
-        statusText: 'OK',
-        cached: false,
-      };
-      vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-      await provider.callApi('test data');
-
-      // Verify URL is sanitized in the log message
-      expect(loggerDebugSpy).toHaveBeenCalledWith(
-        expect.stringContaining('api_key=%5BREDACTED%5D'),
-        expect.anything(),
-      );
-    });
-
-    it('should not impact performance significantly', async () => {
-      const provider = new HttpProvider('https://api.example.com/perf', {
-        config: {
-          method: 'POST',
-          body: { test: 'performance' },
-          headers: {
-            Authorization: 'Bearer perf-token-123',
-          },
-        },
-      });
-
-      const mockResponse = {
-        data: '{"result": "success"}',
-        status: 200,
-        statusText: 'OK',
-        cached: false,
-      };
-      vi.mocked(fetchWithCache).mockResolvedValue(mockResponse);
-
-      const startTime = Date.now();
-
-      // Run multiple calls to test performance impact
-      const promises = Array.from({ length: 5 }, () => provider.callApi('performance test'));
-
-      await Promise.all(promises);
-
-      const endTime = Date.now();
-      const totalTime = endTime - startTime;
-
-      // Should complete reasonably quickly (less than 500ms for 5 calls)
-      expect(totalTime).toBeLessThan(500);
-
-      // Verify logger was called multiple times
-      const debugCalls = loggerDebugSpy.mock.calls.filter(
-        (call: any) => call[0]?.includes('Calling') && call[0]?.includes('with config'),
-      );
-      expect(debugCalls.length).toBeGreaterThan(0);
-    });
-  });
-
-  describe('Edge cases', () => {
-    it('should handle empty or undefined values gracefully', async () => {
-      const provider = new HttpProvider('https://api.example.com/test', {
-        config: {
-          method: 'POST',
-          body: { test: 'value' },
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: '', // Empty header value
-          },
-        },
-      });
-
-      const mockResponse = {
-        data: '{"result": "test"}',
-        status: 200,
-        statusText: 'OK',
-        cached: false,
-      };
-      vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-      // Should not crash
-      await expect(provider.callApi('test prompt')).resolves.not.toThrow();
-
-      // Should still log something
-      expect(loggerDebugSpy).toHaveBeenCalled();
-    });
-
-    it('should handle malformed URLs in sanitizeUrl function', () => {
-      const malformedInputs = ['not-a-url', '', 'https://[invalid-host]/api'];
-
-      malformedInputs.forEach((input) => {
-        expect(() => sanitizeUrl(input)).not.toThrow();
-        const result = sanitizeUrl(input);
-        expect(result).toBeDefined();
-      });
-
-      // Test null/undefined separately as they return the input as-is
-      expect(sanitizeUrl(null as any)).toBeNull();
-      expect(sanitizeUrl(undefined as any)).toBeUndefined();
-    });
-  });
-});
-
-describe('HttpProvider - Abort Signal Handling', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it('should pass abortSignal to fetchWithCache', async () => {
-    const provider = new HttpProvider('http://example.com/api', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: { key: '{{ prompt }}' },
-      },
-    });
-
-    const abortController = new AbortController();
-    const mockResponse = {
-      data: JSON.stringify({ result: 'response text' }),
-      status: 200,
-      statusText: 'OK',
-      cached: false,
-    };
-    vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-    await provider.callApi('test prompt', undefined, { abortSignal: abortController.signal });
-
-    expect(fetchWithCache).toHaveBeenCalledWith(
-      expect.any(String),
-      expect.objectContaining({ signal: abortController.signal }),
-      expect.any(Number),
-      expect.any(String),
-      undefined,
-      undefined,
-    );
-  });
-
-  it('should pass abortSignal to fetchWithCache in raw request mode', async () => {
-    const rawRequest = dedent`
-      POST /api HTTP/1.1
-      Host: example.com
-      Content-Type: application/json
-
-      {"key": "{{ prompt }}"}
-    `;
-
-    const provider = new HttpProvider('http://example.com', {
-      config: {
-        request: rawRequest,
-      },
-    });
-
-    const abortController = new AbortController();
-    const mockResponse = {
-      data: JSON.stringify({ result: 'response text' }),
-      status: 200,
-      statusText: 'OK',
-      cached: false,
-    };
-    vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-    await provider.callApi('test prompt', undefined, { abortSignal: abortController.signal });
-
-    expect(fetchWithCache).toHaveBeenCalledWith(
-      expect.any(String),
-      expect.objectContaining({ signal: abortController.signal }),
-      expect.any(Number),
-      expect.any(String),
-      undefined,
-      undefined,
-    );
-  });
-
-  it('should work without abortSignal (backwards compatibility)', async () => {
-    const provider = new HttpProvider('http://example.com/api', {
-      config: {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: { key: '{{ prompt }}' },
-      },
-    });
-
-    const mockResponse = {
-      data: JSON.stringify({ result: 'response text' }),
-      status: 200,
-      statusText: 'OK',
-      cached: false,
-    };
-    vi.mocked(fetchWithCache).mockResolvedValueOnce(mockResponse);
-
-    // Call without options parameter
-    await provider.callApi('test prompt');
-
-    expect(fetchWithCache).toHaveBeenCalledWith(
-      expect.any(String),
-      expect.not.objectContaining({ signal: expect.anything() }),
-      expect.any(Number),
-      expect.any(String),
-      undefined,
-      undefined,
-    );
-  });
-});
-
 describe('HttpProvider - OAuth Token Refresh Deduplication', () => {
   const mockUrl = 'http://example.com/api';
   const tokenUrl = 'https://auth.example.com/oauth/token';
@@ -1780,8 +594,8 @@ describe('HttpProvider - OAuth Token Refresh Deduplication', () => {
     });
 
     // First call will fail, but subsequent calls should retry
-    const promise1 = provider.callApi('test 1').catch(() => {
-      // Expected to fail
+    const promise1 = provider.callApi('test 1').catch((err: Error) => {
+      expect(err.message).toMatch(/token|auth|refresh|401|fetch/i);
     });
     await firstRefreshStarted.promise;
     // Second call should trigger a retry
@@ -1867,8 +681,8 @@ describe('HttpProvider - OAuth Token Refresh Deduplication', () => {
       return apiResponse;
     });
 
-    const promise1 = provider.callApi('test 1').catch(() => {
-      // Expected to fail on the first refresh attempt.
+    const promise1 = provider.callApi('test 1').catch((err: Error) => {
+      expect(err.message).toMatch(/token|auth|refresh|401|fetch/i);
     });
     await firstRefreshStarted.promise;
     const promise2 = provider.callApi('test 2');
@@ -2238,8 +1052,8 @@ describe('HttpProvider - File Auth', () => {
       },
     });
 
-    const promise1 = provider.callApi('test 1').catch(() => {
-      // Expected to fail on the first refresh attempt.
+    const promise1 = provider.callApi('test 1').catch((err: Error) => {
+      expect(err.message).toMatch(/token|auth|refresh|401|fetch/i);
     });
     await firstRefreshStarted.promise;
     const promise2 = provider.callApi('test 2');
